third_party/WebKit/LayoutTests/intersection-observer/iframe-root.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<script src="helper-functions.js"></script>
<div style="width:100%; height:700px;"></div>
<div id="root" style="display: inline-block; overflow-y: scroll; height: 200px;">
  <div id="target" style="width:100px; height:300px"></div>
  <iframe id="target-iframe" src="../resources/intersection-observer-subframe.html" style="width: 300px; height: 100px; overflow-y: scroll"></iframe>
</div>
<div style="width:100%; height:700px;"></div>

<script>
description("Simple intersection observer test with no explicit root and target in an iframe.");
var entries = [];
var root = document.getElementById("root");
var targetIframe = document.getElementById("target-iframe");

targetIframe.onload = function() {
  var target = targetIframe.contentDocument.getElementById("target");
  var iframeScroller = targetIframe.contentDocument.scrollingElement;

  observer_callback = function(changes) {
    for (var i in changes)
      entries.push(changes[i]);
  };
  var observer = new IntersectionObserver(observer_callback, {"root": document.getElementById("root")});
  observer.observe(target);

  function step0() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 0);
      document.scrollingElement.scrollTop = 250;
      requestAnimationFrame(step1);
    });
  }

  function step1() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 0);
      iframeScroller.scrollTop = 200;
      requestAnimationFrame(step2);
    });
  }

  function step2() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 0);
      root.scrollTop = 175;
      requestAnimationFrame(step3);
    });
  }

  function step3() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 1);
      shouldBeEqualToNumber("entries[0].boundingClientRect.left", 18);
      shouldBeEqualToNumber("entries[0].boundingClientRect.right", 118);
      shouldBeEqualToNumber("entries[0].boundingClientRect.top", 593);
      shouldBeEqualToNumber("entries[0].boundingClientRect.bottom", 693);
      shouldBeEqualToNumber("entries[0].intersectionRect.left", 18);
      shouldBeEqualToNumber("entries[0].intersectionRect.right", 118);
      shouldBeEqualToNumber("entries[0].intersectionRect.top", 593);
      shouldBeEqualToNumber("entries[0].intersectionRect.bottom", 658);
      shouldBeEqualToNumber("entries[0].rootBounds.left", 8);
      shouldBeEqualToNumber("entries[0].rootBounds.right", 312);
      shouldBeEqualToNumber("entries[0].rootBounds.top", 458);
      shouldBeEqualToNumber("entries[0].rootBounds.bottom", 658);
      shouldEvaluateToSameObject("entries[0].target", target);
      document.scrollingElement.scrollTop = 100;
      requestAnimationFrame(step4);
    });
  }

  function step4() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 1);
      root.scrollTop = 100;
      requestAnimationFrame(step5);
    });
  }

  function step5() {
    setTimeout(function() {
      shouldBeEqualToNumber("entries.length", 2);
      shouldBeEqualToNumber("entries[1].boundingClientRect.left", 18);
      shouldBeEqualToNumber("entries[1].boundingClientRect.right", 118);
      shouldBeEqualToNumber("entries[1].boundingClientRect.top", 818);
      shouldBeEqualToNumber("entries[1].boundingClientRect.bottom", 918);
      shouldBeEqualToNumber("entries[1].intersectionRect.left", 0);
      shouldBeEqualToNumber("entries[1].intersectionRect.right", 0);
      shouldBeEqualToNumber("entries[1].intersectionRect.top", 0);
      shouldBeEqualToNumber("entries[1].intersectionRect.bottom", 0);
      shouldBeEqualToNumber("entries[1].rootBounds.left", 8);
      shouldBeEqualToNumber("entries[1].rootBounds.right", 312);
      shouldBeEqualToNumber("entries[1].rootBounds.top", 608);
      shouldBeEqualToNumber("entries[1].rootBounds.bottom", 808);
      shouldEvaluateToSameObject("entries[1].target", target);
      finishJSTest();
      document.scrollingElement.scrollTop = 0;
    });
  }

  step0();
}
</script>
