third_party/WebKit/LayoutTests/intersection-observer/iframe-root-expected.txt

Simple intersection observer test with no explicit root and target in an iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 0
PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 18
PASS entries[0].boundingClientRect.right is 118
PASS entries[0].boundingClientRect.top is 593
PASS entries[0].boundingClientRect.bottom is 693
PASS entries[0].intersectionRect.left is 18
PASS entries[0].intersectionRect.right is 118
PASS entries[0].intersectionRect.top is 593
PASS entries[0].intersectionRect.bottom is 658
PASS entries[0].rootBounds.left is 8
PASS entries[0].rootBounds.right is 312
PASS entries[0].rootBounds.top is 458
PASS entries[0].rootBounds.bottom is 658
PASS entries[0].target is [object HTMLDivElement]
PASS entries.length is 1
PASS entries.length is 2
PASS entries[1].boundingClientRect.left is 18
PASS entries[1].boundingClientRect.right is 118
PASS entries[1].boundingClientRect.top is 818
PASS entries[1].boundingClientRect.bottom is 918
PASS entries[1].intersectionRect.left is 0
PASS entries[1].intersectionRect.right is 0
PASS entries[1].intersectionRect.top is 0
PASS entries[1].intersectionRect.bottom is 0
PASS entries[1].rootBounds.left is 8
PASS entries[1].rootBounds.right is 312
PASS entries[1].rootBounds.top is 608
PASS entries[1].rootBounds.bottom is 808
PASS entries[1].target is [object HTMLDivElement]
PASS successfullyParsed is true

TEST COMPLETE
